requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
bitsandbytes==0.42.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
comm==0.2.2
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
evaluate==0.4.3
executing==2.2.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.3
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.10
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
openai==1.85.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pfzy==0.3.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
submitit==1.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
trl==0.18.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wandb==0.19.11
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
